Implemented data access to the feedback services in the View Model. Updated 2sxc model mapping to use a double for order instead of an int

+++ src/DnnSummit.Data/TwoSexyContent/Feedback.cs
@@ -15,6 +15,6 @@ namespace DnnSummit.Data.TwoSexyContent
         public string Type { get; set; }
 
         [JsonProperty]
-        public int Order { get; set; }
+        public double Order { get; set; }
     }
 }
